refactor: reorganize test package structure and update imports
