+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/AbstractUITest.java
@@ -6,6 +6,7 @@ import org.openqa.selenium.WebDriver;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import com.hack23.cia.systemintegrationtest.suites.TestConstants;
 import com.hack23.cia.systemintegrationtest.ui.UserPageVisit;
 import com.hack23.cia.systemintegrationtest.ui.WebDriverFactory;
