+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/admin/AdminAgentTest.java
@@ -1,4 +1,4 @@
-package com.hack23.cia.systemintegrationtest;
+package com.hack23.cia.systemintegrationtest.admin;
 
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
@@ -6,7 +6,9 @@ import org.junit.runner.RunWith;
 import org.junit.runners.Parameterized;
 import org.openqa.selenium.WebElement;
 
-import com.hack23.cia.systemintegrationtest.suites.IntegrationTest;
+import com.hack23.cia.systemintegrationtest.AbstractUITest;
+import com.hack23.cia.systemintegrationtest.categories.IntegrationTest;
+import com.hack23.cia.systemintegrationtest.suites.TestConstants;
 import com.hack23.cia.web.impl.ui.application.views.common.pagelinks.api.PageModeMenuCommand;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.AdminViews;
 
@@ -15,7 +17,7 @@ import com.hack23.cia.web.impl.ui.application.views.common.viewnames.AdminViews;
 public final class AdminAgentTest extends AbstractUITest {
 
 
-    @Test(timeout = DEFAULT_TIMEOUT, expected = Exception.class)
+    @Test(timeout = TestConstants.DEFAULT_TIMEOUT, expected = Exception.class)
     public void shouldStartImportOperations() throws Exception {
         try {
             retryOnFailure(() -> {
