+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/admin/AdminRoleSystemITest.java
@@ -16,17 +16,17 @@
  *	$Id$
  *  $HeadURL$
 */
-package com.hack23.cia.systemintegrationtest;
+package com.hack23.cia.systemintegrationtest.admin;
 
-import static com.hack23.cia.systemintegrationtest.TestConstants.DEFAULT_MAX_RETRIES;
+import static com.hack23.cia.systemintegrationtest.suites.TestConstants.DEFAULT_MAX_RETRIES;
 
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
-import org.junit.runner.RunWith;
-import org.junit.runners.Parameterized;
 import org.openqa.selenium.WebElement;
 
-import com.hack23.cia.systemintegrationtest.suites.IntegrationTest;
+import com.hack23.cia.systemintegrationtest.AbstractUITest;
+import com.hack23.cia.systemintegrationtest.categories.IntegrationTest;
+import com.hack23.cia.systemintegrationtest.suites.TestConstants;
 
 // import com.hack23.cia.systemintegrationtest.TestConstants; // Unused import removed
 // import com.hack23.cia.web.impl.ui.application.action.ViewAction; // Unused import removed
