+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/admin/SessionManagementTest.java
@@ -3,10 +3,10 @@ package com.hack23.cia.systemintegrationtest.admin;
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
 import org.openqa.selenium.WebElement;
-import com.hack23.cia.systemintegrationtest.TestConstants;
 
-
-import com.hack23.cia.systemintegrationtest.suites.IntegrationTest;
+import com.hack23.cia.systemintegrationtest.AbstractUITest;
+import com.hack23.cia.systemintegrationtest.categories.IntegrationTest;
+import com.hack23.cia.systemintegrationtest.suites.TestConstants;
 import com.hack23.cia.web.impl.ui.application.views.common.pagelinks.api.PageModeMenuCommand;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.AdminViews;
