+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/user/ChartTest.java
@@ -16,7 +16,7 @@
  *	$Id$
  *  $HeadURL$
 */
-package com.hack23.cia.systemintegrationtest;
+package com.hack23.cia.systemintegrationtest.user;
 
 import java.util.Map;
 import java.util.Map.Entry;
