+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/user/SearchViewTest.java
@@ -1,9 +1,11 @@
 package com.hack23.cia.systemintegrationtest.user;
 
+import static com.hack23.cia.systemintegrationtest.suites.TestConstants.DEFAULT_MAX_RETRIES;
+
 import org.junit.Test;
 import org.junit.experimental.categories.Category;
-import static com.hack23.cia.systemintegrationtest.TestConstants.DEFAULT_MAX_RETRIES;
 
+import com.hack23.cia.systemintegrationtest.AbstractUITest;
 import com.hack23.cia.systemintegrationtest.categories.IntegrationTest;
 import com.hack23.cia.web.impl.ui.application.views.common.pagelinks.api.PageModeMenuCommand;
 import com.hack23.cia.web.impl.ui.application.views.common.viewnames.UserViews;
