+++ citizen-intelligence-agency/src/test/java/com/hack23/cia/systemintegrationtest/user/package-info.java
@@ -1,4 +1,4 @@
 /**
  * Provides...
  */
-package com.hack23.cia.systemintegrationtest;
+package com.hack23.cia.systemintegrationtest.user;
